### 2018-10-24 #1048 @whedon
[PRE REVIEW]: TextWiller: collection of text mining utilities, specially devoted to the Italian language
**Submitting author:** @scan2001 (<a href="http://orcid.org/0000-0001-5621-5463">andrea sciandra</a>)
**Repository:** <a href="https://github.com/livioivil/TextWiller" target ="_blank">https://github.com/livioivil/TextWiller</a>
**Version:** 2.0
