### 2018-11-19 #1048 @arfon
:wave: @mgymrek - would you be able to edit this submission for JOSS?
