### 2018-11-23 07f5df4
Update paper.md
