### 2018-11-23 #1048 @scan2001
@labarba : thank you for your advice. I edited the paper with some sentences about the the research application of the software and the targeted audience.
