### 2018-12-03 #1048 @danielskatz
👋 @mgymrek - would you be able to edit this submission for JOSS? (reposting)
