### 2019-01-21 #1048 @mgymrek
@linuxscout @timClicks @lmullen would any of you be able to review this submission?
