### 2019-02-11 #1048 @livioivil
> **Authorship**

> I believe that it's up to your team to decide on authorship, but the three authors don't match the authors listed in the [DESCRIPTION](https://github.com/livioivil/TextWiller/blob/master/DESCRIPTION) file or the [TextWiller-package.R](https://github.com/livioivil/TextWiller/blob/master/R/TextWiller-package.R#L13) file.

> Also, from the [contribution list](https://github.com/livioivil/TextWiller/graphs/contributors), it's unclear what level of contribution that your first author has made? (I assume that some of the commit history has been lost)

Hi, I'm the host of the package. I confirm that I have done a large part of the submissions. However, @scan2001 in most of the cases scripted the first version of the code. So, I think, he deserves the merit of the first authorship.
I hope this helps.
