### 2019-02-15 #1256 @mgymrek
Copying from the pre-review thread from @timClicks :

Hi @scan2001, thanks for the paper submission. I'm not officially an assigned reviewer, but I've taken a quick look at the paper.

Let me start by saying that it's excellent to see NLP contributions to non-English languages.

**Prior art**

However, there are many text mining packages available, even sticking within the R community. I wonder if you should cite some of the other open source NLP packages and explain that they do not offer Italian stemming/normalisation?

From the review guidelines:

> Submissions that implement solutions already solved in other software packages are accepted into JOSS provided that they meet the criteria listed above **and cite prior similar work**. [emphasis added]

**Authorship**

I believe that it's up to your team to decide on authorship, but the three authors don't match the authors listed in the [DESCRIPTION](https://github.com/livioivil/TextWiller/blob/master/DESCRIPTION) file or the [TextWiller-package.R](https://github.com/livioivil/TextWiller/blob/master/R/TextWiller-package.R#L13) file.

Also, from the [contribution list](https://github.com/livioivil/TextWiller/graphs/contributors), it's unclear what level of contribution that your first author has made? (I assume that some of the commit history has been lost)

Here is the relevant guidance from the review guidelines:

> As part of the review process, you are asked to check whether the submitting author has made a ‘substantial contribution’ to the submitted software (as determined by the commit history) and to check that ‘the full list of paper authors seems appropriate and complete?’

**Documentation**

[TextWiller's API documentation](https://github.com/livioivil/TextWiller/tree/master/man) is in Italian. However, it looks comprehensive.

(Deferring to @mgymrek for an opinion about whether this is allowed)

**Tests**

TextWiller doesn't appear to have a comprehensive test suite.
