### 2019-05-27 #1256 @timClicks
Statement of need from paper.md, l 32

> The main quality of this software is to be one of the few text mining R packages in Italian language. Moreover, TextWiller can help social media researchers with some specific functions for the data extracted from Twitter via APIs.
