### 2019-05-27 #1256 @timClicks
Installation was very easy, `TextWiller`'s dependencies were fully satisfied by installing the `devtools` library as recommended in the instructions.
