### 2019-05-27 #1256 @timClicks
**Functionality documentation** provided by `TextWiller` is comprehensive with examples. All public functions appear to be documented in RMarkdown in the `/man` directory, as is conventional for R projects.
