### 2019-05-27 #1256 @timClicks
@mgymrek I believe that there are no community guidelines in the repository or any tests, either automated or manual.

I believe that the paper and the project meet the other criteria.
