### 2019-06-06 #1256 @mgymrek
Thanks @timClicks.

@scan2001. I agree with these comments. Including tests and community guidelines is a JOSS requirement. Are you able to add tests for TextWiller functionality, and also to expand the README with guidelines for community members to contribute or get support?

You may look at other JOSS submissions for examples. e.g. https://github.com/ropensci/citesdb
