### 2019-06-07 #1256 @scan2001
Thanks @mgymrek and @timClicks.
We will add tests for TextWiller functionality and the guidelines for community members to contribute or get support.
