### 2019-06-26 #1256 @kyleniemeyer
Hi @scan2001, just wanted to check on your progress. Can you give us an expected date when you think you'll be able to complete that work?
