### 2019-06-26 #1256 @scan2001
Hi @kyleniemeyer, sorry for the delay. We have already included the guidelines for community members and we will run the tests probably next week, with maximum expected date no later than July 14th.
