+++ man/urlExtract.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/urlExtract.R
 \name{urlExtract}
-\alias{patternExtract}
-\alias{shorturl2url}
 \alias{urlExtract}
+\alias{shorturl2url}
+\alias{patternExtract}
 \title{Estrazione di regular expression (e quindi users, hashtag) e shorturl}
 \usage{
 urlExtract(testo, id = names(testo))
@@ -49,4 +49,3 @@ Dario Solari, Livio Finos
 }
 \keyword{~kwd1}
 \keyword{~kwd2}
-
