+++ man/sentiment.Rd
@@ -37,4 +37,3 @@ Maddalena Branca, Livio Finos
 }
 \keyword{~kwd1}
 \keyword{~kwd2}
-
