+++ man/fixTimeStamp.Rd
@@ -41,4 +41,3 @@ Dario Solari, Livio Finos
 }
 \keyword{~kwd1}
 \keyword{~kwd2}
-
