+++ man/normalizzaTesti.Rd
@@ -1,25 +1,25 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/normalizzaTesti.R
 \name{normalizzaTesti}
-\alias{itastopwords}
 \alias{normalizzaTesti}
 \alias{normalizzacaratteri}
 \alias{normalizzaemote}
 \alias{normalizzahtml}
-\alias{normalizzapunteggiatura}
 \alias{normalizzaslang}
-\alias{preprocessingEncoding}
-\alias{removeStopwords}
+\alias{normalizzapunteggiatura}
 \alias{tryTolower}
+\alias{itastopwords}
+\alias{removeStopwords}
+\alias{preprocessingEncoding}
 \title{Varie funzioni di normalizzazione del testo}
 \usage{
 normalizzaTesti(testo, tolower = TRUE, normalizzahtml = TRUE,
   normalizzacaratteri = TRUE, normalizzaemote = TRUE,
-  normalizzapunteggiatura = TRUE, normalizzaslang = TRUE, fixed = TRUE,
-  perl = TRUE, preprocessingEncoding = TRUE, encoding = "UTF-8",
-  sub = "", contaStringhe = c("\\\\?", "\\\\!", "@", "#", "(€|euro)",
-  "(\\\\$|dollar)", "SUPPRESSEDTEXT"), suppressInvalidTexts = TRUE,
-  verbatim = TRUE, remove = TRUE)
+  normalizzapunteggiatura = TRUE, normalizzaslang = TRUE,
+  fixed = TRUE, perl = TRUE, preprocessingEncoding = TRUE,
+  encoding = "UTF-8", sub = "", contaStringhe = c("\\\\?", "\\\\!",
+  "@", "#", "(\\200|euro)", "(\\\\$|dollar)", "SUPPRESSEDTEXT"),
+  suppressInvalidTexts = TRUE, verbatim = TRUE, remove = TRUE)
 }
 \arguments{
 \item{testo}{character vector of texts}
@@ -100,4 +100,3 @@ Dario Solari, Livio Finos, Maddalena Branca
 }
 \keyword{~kwd1}
 \keyword{~kwd2}
-
