+++ DESCRIPTION
@@ -17,4 +17,7 @@ Depends:
 Description: Collection of functions for text mining, specially devoted to the
     italian language.
 License: GPL (>= 2)
-RoxygenNote: 5.0.1
+RoxygenNote: 6.1.1
+Suggests: 
+    testthat
+Encoding: UTF-8
