### 2019-06-28 #1256 @scan2001
@mgymrek @kyleniemeyer
we have completed the tests and added community guidelines (updating README and creating a code of conduct file)
