### 2019-07-16 #1256 @mgymrek
Thanks for the updates. Comments:

* Please add to the README how to run the tests. This wasn't clear.

For the paper:

* The paper should mention in the first sentence the focus on Italian language.
* The paper should have a clear statement of need: are there any other existing packages that do something similar? How does your tool compare to those?
* The text could be slighltly expanded so non-experts can better interpret what the tool is doing. For example, you might have bullet points or small paragraph about several of the major functions, where could you describe and define the goal (e.g. what does it mean to normalize) and the inputs and outputs of the tool.
