+++ README.md
@@ -140,4 +140,6 @@ patternExtract(c("@luca @paolo: buon giorno!", "@matteo: a te!"), pattern="@\\w+
 
 You can contribute by reporting an issue or by directly contributing to the source code. You can create a fork of the repo in your own GitHub account and download the repository to your local machine.
 
-Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms. Please include tests with any major contribution to code. Test your contributions with [**testthat**](https://cran.r-project.org/web/packages/testthat/index.html) before submitting your change.
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms. 
+
+Please include tests with any major contribution to code. Test your contributions with [**testthat**](https://cran.r-project.org/web/packages/testthat/index.html) before submitting your change.
