### 2019-07-29 868b001
Update paper.md
