### 2019-07-30 #1256 @scan2001
@mgymrek we modified the paper and the README to meet your requests
