### 2019-08-04 #1256 @mgymrek
Great, thanks for the changes. This looks good.

To prepare for publication can you:

* Fix the following minor typos:
"few text mining R packages in Italian language" -> "few text mining R packages in the Italian language"
"main differences ... is... " -> "main differences ...are..."
"extract users communication pattern" -> "extract users' communication patterns"

* Please add DOIs to all references

* Create an archive (e.g. Zenodo) and post the archive DOI in the review issue
