+++ paper.bib
@@ -1,25 +1,28 @@
-@inproceedings{wilson2005recognizing,
-  title={Recognizing contextual polarity in phrase-level sentiment analysis},
-  author={Wilson, Theresa and Wiebe, Janyce and Hoffmann, Paul},
-  booktitle={Proceedings of the conference on human language technology and empirical methods in natural language processing},
-  pages={347--354},
-  year={2005},
-  organization={Association for Computational Linguistics}
+@inproceedings{Wilson_2005,
+	doi = {10.3115/1220575.1220619},
+	url = {https://doi.org/10.3115%2F1220575.1220619},
+	year = 2005,
+	publisher = {Association for Computational Linguistics},
+	author = {Theresa Wilson and Janyce Wiebe and Paul Hoffmann},
+	title = {Recognizing contextual polarity in phrase-level sentiment analysis},
+	booktitle = {Proceedings of the conference on Human Language Technology and Empirical Methods in Natural Language Processing  - {HLT} {\textquotesingle}05}
 }
 
-@book{ceron2014social,
-  title={Social Media e Sentiment Analysis: L'evoluzione dei fenomeni sociali attraverso la Rete},
-  author={Ceron, Andrea and Curini, Luigi and Iacus, Stefano Maria},
-  volume={9},
-  year={2014},
-  publisher={Springer Science \& Business Media}
+@book{Ceron_2014,
+	doi = {10.1007/978-88-470-5532-2},
+	url = {https://doi.org/10.1007%2F978-88-470-5532-2},
+	year = 2014,
+	publisher = {Springer Milan},
+	author = {Andrea Ceron and Luigi Curini and Stefano M. Iacus},
+	title = {Social Media e Sentiment Analysis}
 }
 
-@book{miner2012practical,
-  title={Practical text mining and statistical analysis for non-structured text data applications},
-  author={Miner, Gary and Elder IV, John and Hill, Thomas},
-  year={2012},
-  publisher={Academic Press}
+@book{2012,
+	doi = {10.1016/c2010-0-66188-8},
+	url = {https://doi.org/10.1016%2Fc2010-0-66188-8},
+	year = 2012,
+	publisher = {Elsevier},
+	title = {Practical Text Mining and Statistical Analysis for Non-structured Text Data Applications}
 }
 
 @book{bolasco2013analisi,
