+++ paper.md
@@ -23,7 +23,7 @@ affiliations:
    index: 2
  - name: Department of Developmental Psychology and Socialisation, University of Padova
    index: 3
-date: 29 July 2019
+date: 20 August 2019
 bibliography: paper.bib
 ---
 
