### 2019-08-26 c6ea91c
Update paper.md
