### 2019-08-26 e7fbda0
Update paper.bib
