+++ paper.bib
@@ -17,9 +17,10 @@
 	title = {Social Media e Sentiment Analysis}
 }
 
-@book{2012,
+@book{miner2012practical,
 	doi = {10.1016/c2010-0-66188-8},
 	url = {https://doi.org/10.1016%2Fc2010-0-66188-8},
+	author={Miner, Gary and Elder IV, John and Fast, Andrew and Hill, Thomas and Nisbet, Robert and Delen, Dursun},
 	year = 2012,
 	publisher = {Elsevier},
 	title = {Practical Text Mining and Statistical Analysis for Non-structured Text Data Applications}
