### 2019-08-30 #1256 @scan2001
Thanks @mgymrek
We fixed the typos, added DOIs to all references (except the last two: a book and a thesis, as they haven't a DOI) and created an archive in Zenodo.
This is the archive DOI: 10.5281/zenodo.3381523
