### 2019-09-01 #1256 @mgymrek
Thanks @scan2001. Can you please ensure the Zenodo release has the same title and author list as the JOSS paper for consistency? Then we should be ready to go.
