+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: TextWiller
 Type: Package
-Title: Collection of functions for text mining, specially devoted to the italian
+Title: Collection of functions for text mining, specially devoted to the Italian
     language
 Version: 2.0
 Date: 2016-05-1
