### 2019-09-02 #1256 @scan2001
Thanks @mgymrek. We set Zenodo release title and authors so that they are the same as those in the paper
