### 2019-09-04 #1256 @mgymrek
Thanks @scan2001. I noticed the version on this thread is listed as 2.0 but on Zenodo it says v1.0. Could you clarify which version number the final release is, and make sure that matches with what is listed on Zenodo?
