### 2019-09-04 #1256 @scan2001
Thanks @mgymrek. We named the release v1.0 because it's the first release we did specifically for the Zenodo archive, while version 2.0 doesn't concern the release, it was only written by one collaborator in the description without any particular reason, actually we don't have 2 version of this software.
